pyproject.toml
[tool.poetry]
name = "ChatGPT-API-Recipes"
version = "0.1.0"
description = ""
authors = ["perforb <dev.perfumed.garden@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.3.3"
python-dotenv = "^1.0.0"
tweepy = "^4.14.0"
pandas = "^2.1.3"
tiktoken = "^0.5.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
